feat(dashboard): 부서별 통계 배치 저장 로직 및 DTO 타입 수정

+++ src/main/java/io/project/backend/domain/employee/dto/common/DashboardDepartmentStatsDto.java
@@ -5,7 +5,7 @@ import java.time.LocalDate;
 public record DashboardDepartmentStatsDto(
     LocalDate snapshotDate,
     String departmentName,
-    Long employeeCount,
+    int employeeCount,
     String managerName
 ) {
 
@@ -19,7 +19,7 @@ public record DashboardDepartmentStatsDto(
     return new DashboardDepartmentStatsDto(
         now,
         departmentName,
-        employeeCount,
+        employeeCount.intValue(),
         managerName
     );
   }
